Add files via upload
Adding a Spectral Extractor that uses the CCA algorithm for SSVEP applications. In settings, we define a number of Frequencies of Interest that we want to survey for in an incoming data stream. The SpectraCcaExtractor unit then checks the incoming stream for those frequencies of interest and returns the strongest one.